Map Screen

+++ app/src/main/java/com/aajogo/jogo/photosapp/data/models/data/ImageResponse.kt
@@ -6,6 +6,6 @@ data class ImageResponse(
     @Json(name = "id") val id: Int? = null,
     @Json(name = "url") val url: String? = null,
     @Json(name = "date") val date: Int? = null,
-    @Json(name = "lat") val lat: Int? = null,
-    @Json(name = "lng") val lng: Int? = null,
+    @Json(name = "lat") val lat: Double? = null,
+    @Json(name = "lng") val lng: Double? = null,
 )
